Move overrides to a separate module
