+++ utils/model/gen_uml.py
@@ -98,6 +98,7 @@ class Writer:
 
         try:
             out.write(header)
+            out.write(self.overrides.header)
             for cls in self.classdefs.values():
                 if cls[0].endswith(":"):
                     for d in cls:
